requirements.txt
airtest==1.3.2
loguru==0.7.2
matplotlib==3.2.1
numpy==1.18.4
onnx==1.15.0
onnxruntime==1.16.3
onnxsim==0.4.35
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
paddleocr==2.7.0.2
paddleocr.egg==info
Pillow==9.5.0
Pillow==10.1.0
PyAutoGUI==0.9.54
pycocotools==2.0.7
pyserial==3.5
pywin32==306
scipy==1.4.1
tensorflow_gpu==2.3.0
tf2onnx==1.15.1
tqdm==4.46.1
